+++ packages/moonstone/CHANGELOG.md
@@ -6,11 +6,11 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Fixed
 
-- `moonstone/Input` text color
 - `moonstone` internationalization resource loading
+- `moonstone/Dropdown` to only call `onOpen` when closed
+- `moonstone/Input` text color
 - `moonstone/VirtualList.VirtualGridList` and `moonstone/VirtualList.VirtualList` scrolling when an item which will be in viewport gets focus
 - `moonstone/VirtualList.VirtualList` to scroll properly when an item gets focus in VirtualList with different item size
-- `moonstone/Dropdown` to call `onOpen` handler once
 
 ## [3.1.0] - 2019-09-16